requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
python-dotenv
pydantic-settings
praw
pandas
vaderSentiment
APScheduler
